Rename index.js (the server) -> server.js; exports.js -> index.js; make browserify the default Grunt task
