+++ Gruntfile.js
@@ -2,10 +2,10 @@ module.exports = function(grunt) {
     grunt.initConfig({
 	"pkg": grunt.file.readJSON('package.json'),
 	"browserify": {
-	    "./public/index.js": ["./exports.js"],
+	    "./public/index.js": ["./index.js"],
 	    "options": {
 		"browserifyOptions": {
-		    "require": "./exports.js",
+		    "require": "./index.js",
 		    "outfile": "./public/index.js",
 		    "standalone": "DattLib"
 		}
@@ -18,5 +18,6 @@ module.exports = function(grunt) {
     });
 
     grunt.loadNpmTasks("grunt-browserify");
-    grunt.loadNpmTasks('grunt-contrib-watch');
+    grunt.loadNpmTasks("grunt-contrib-watch");
+    grunt.registerTask("default", "browserify");
 };
