--- exports.js
@@ -1,16 +1,0 @@
-var bitcoin = require('bitcoinjs-lib');
-var Peer = require('peerjs');
-
-var Content = require('./lib/content');
-var User = require('./lib/user');
-var Message = require('./lib/message');
-var Datt = require('./lib/datt');
-
-module.exports = {
-    "Content": Content,
-    "User": User,
-    "Message": Message,
-    "Datt": Datt,
-    "Peer": Peer,
-    "bitcoin": bitcoin
-};
