+++ index.js
@@ -1,23 +1,16 @@
-var peer = require('peer');
-var PeerServer = peer.PeerServer;
+var bitcoin = require('bitcoinjs-lib');
+var Peer = require('peerjs');
 
-var peerServerConfig = require('./config/peerServer.js');
+var Content = require('./lib/content');
+var User = require('./lib/user');
+var Message = require('./lib/message');
+var Datt = require('./lib/datt');
 
-var peerServer = new PeerServer(peerServerConfig);
-
-console.log("PeerServer started, details below:");
-console.log(JSON.stringify(peerServerConfig, null, 4));
-
-var express = require('express');
-
-var app = express();
-
-app.use(express.static('public'));
-
-var server = app.listen(3000, function () {
-  var host = server.address().address;
-  var port = server.address().port;
-
-  console.log('Static content at http://%s:%s', host, port);
-});
-
+module.exports = {
+    "Content": Content,
+    "User": User,
+    "Message": Message,
+    "Datt": Datt,
+    "Peer": Peer,
+    "bitcoin": bitcoin
+};
